requirements.txt
aiohttp
azure-cosmos
cryptography
fastapi
httpx
itsdangerous
langchain-core
langchain-mcp-adapters
langchain-openai
langgraph
langgraph-cli[inmem]
langgraph-sdk
langsmith
msal
pydantic
python-dotenv
pyjwt
starlette
supabase